Add example after intro
As discussed with @sebmarkbage we should have a concrete example at the top of the page so people can get an idea of what JSX could be used for.
While I understand that this page is primarily a spec, it still could end up ranking high in search and would be pretty bad marketing if people bounced more confused than they came in.