Migrate to ecmarkup
This is the first step to revamp this repo: [ecmarkup](https://github.com/tc39/ecmarkup)
is TC39's official tool for specifying syntax and semantics for ECMAScript.
I proposed that we should adopt it because it has several favorabilities
than the current approach (`README.md` + a `gh-pages` website living in
the `websitescript` branch):
1. The resultant `index.html` has a similar looks and feels with the
   official ECMA-262 spec.
2. It uses [Grammarkdown](https://github.com/rbuckton/grammarkdown)
   which is not only convenient to write but more importantly, can
   enforce the grammar to be more formal.
3. It provides links to ECMA-262 out of the box (closed #112).
   Furthurmore, I can use `<ins>` to make "extension" points explicit.

One possible concern it that it'll make this too like a ECMAScript proposal,
but I think we mediated by opting-out from the "stage-N proposal"
heading, and by explicitly calling out our intention in the
"Introduction" section.

In addition,
- Link to JXON is updated since the original link no more existed.
- Parser and Transpiler section is dropped from the "spec" website.
- Facebook, Inc. is updated to Meta Platform, Inc everywhere.

Once this is merged, we can switch to `main` to host Github Pages.