[Normative] Capture the HTML entity behaviors
Let's be faithful to the de-facto and capture the HTML semantics
to the spec. I haven't seen any practices specifying transipilers
via ECMA-262 so this was a bit challenging may seems foreign.

I ended up extending `Static Semantics: SV` to make it concise and
(hopefully) accurate enough. I think this is a cool hack ;)

- Introduced `JSXStringCharacter` and `JSXString`
- Fix #133 by using `::` to make problematic grammars lexcial

I intentionally making the set of supported HTML entities
implementation-defined to allow either HTML4 or HTML5 set but
this is open to discuss.